Drop preload code for side windows, add comments, rename vars
Also add a "reset" event to drop all 3D effects before switching to
Wikipedia mode.

And make buttons slightly bigger because new raycaster mode makes it slightly
harder to click on them for some reason.

Not clear why performances seem worse than in previous code. On startup, the 5
loading of `about:blank` may perhaps delay things slightly, but the planes are
inert after that, and aren't complex geometry.

That could be due to the raycaster. But why?