Add experiments with wikipedia pages (#1)
* Set up several browser windows for viewing a wikipedia page based on most linked wikipedia pages it links to

* Add support for image companion mode heavily based on wikipedia heuristics

* Handle clicks on main browser window

* Use dynamic values for calculating click positions

* Drop preload code for side windows, add comments, rename vars

Also add a "reset" event to drop all 3D effects before switching to
Wikipedia mode.

And make buttons slightly bigger because new raycaster mode makes it slightly
harder to click on them for some reason.

Not clear why performances seem worse than in previous code. On startup, the 5
loading of `about:blank` may perhaps delay things slightly, but the planes are
inert after that, and aren't complex geometry.

That could be due to the raycaster. But why?