Add default 3D rules for arbitrary pages
Various improvements to 3D mode:
- 3D mode can be enabled on any page, it applies a very basic set of rules
by default. To my surprise, that works well with Wikipedia pages, even though
there are some artefacts. Also, scrolling is *not* supported (not to mention
interactions with created 3D planes).
- 3D toggle animates the planes in and out for smoother transitions. That's
not perfect but, hey, that's a proof of concept!
- Enabling 3D no longer requires a 'paint' event. The code currently keeps an
eye on the last painted image so that it may extract the feature rectangles
when needed. That's probably not a good idea but, hey, did I mention it's a
proof of concept?

Update includes some message renames to make things hopefully more readable.
3D painting is now completely separated from the regular 'paint' event for
instance.

Also, preload.js now also exports a `logger` instance to the immersive
browser so as not to have to open a dev tools when debugging problems.