Add home page and (rough) support for background images
For images, the code no longer takes a screenshot of the rendered content but
rather uses the image as texture directly. This allows to send background
images to the back (the screeshot approach would also contain all content
rendered on top of the image), and seems a better approach for most images.

Now, if CSS is applied to the image to distort its rendering somehow, the
code just won't see it, and the 3D effect won't work as expected.

The image companion and navigator modes can now apply to any page. Default
rule is to look at all images and all links, which will likely not create
a fantastic result on most pages... But it works well enough for demo purpose!

The old index.html page was renamed to browser.html. The home.html page is
now the home page of the experience. It features a background image (any
resemblance with a photo taken during summer vacations would be purely
coincidental...), and a short menu with a set of links. If a URL was given
on the command line, it gets added to the list.

The home page is also great to illustrate transparency and sending the
background image to the back in 3D mode (although note the artefact on the
title, which gets captured with its background).